pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "redshift-streaming-analytics"
version = "0.1.0"
description = "Streaming analytics demo for Redshift query metrics"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "DE Project Team"}]

dependencies = [
  "duckdb>=1.0.0",
  "kafka-python>=2.0.2",
  "pandas>=2.2.0",
  "pyarrow>=14.0.0",
  "pyyaml>=6.0.1",
  "python-dotenv>=1.0.1",
  "streamlit>=1.35.0",
]

[project.optional-dependencies]
redshift = ["boto3>=1.34.0", "psycopg2-binary>=2.9.9"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
